FPS graph

+++ Src/Planet-Engine/Public/Compute/ComputeShader.h
@@ -31,13 +31,13 @@ class ComputeShader
     inline std::string GetShaderName() const { return shader_name_; }
     inline NumThreads GetNumThreads() const { return num_threads_; }
 
-    inline int GetNumTextureOutputs() const { return texture_outputs_.size(); }
+    inline int GetNumTextureOutputs() const { return static_cast<int>(texture_outputs_.size()); }
     inline const Texture* GetTextureOutput(int slot) const { return texture_outputs_[slot].get(); }
 
     void AddTextureOutput(const std::shared_ptr<ComputeTexture2D>& texture);
     void AddTextureOutput(const std::shared_ptr<ComputeTexture3D>& texture);
 
-    inline int GetNumDataInputs() const { return data_outputs_.size(); }
+    inline int GetNumDataInputs() const { return static_cast<int>(data_outputs_.size()); }
     inline const DataBuffer* GetDataInput(int slot) const { return data_outputs_[slot].get(); }
     void AddDataInput(const std::shared_ptr<DataBuffer>& data);
